All user facing strings moved to a common strings file.
Instruction:

"move all user facing strings in the files to a common file having a key and corresponding string to be shown in the view files. Use those keys only in the code  view files. Ensure that the keys are unique for unique strings and are concise and represnt the string appropriately.
"