Added api demo section
Prompt:"
add a section for API demo showing two boxes. One box will contain the api call with payload as in screenshot and another box will be output box and will be prepopulated with the output in screenshot.
User can modify the payload as required and then on pressing 'Try it' cta,  api should be called and return the appropriate output in the output box.
Place the boxes side by side.

Add this section just above the pricing section.
"