Update APIs to REST APIs to be called from client
Update all CRUD apis to REST so that they can be called from client and run on server.
Ensure that apis deal with api keys for the current authenticated user only.

Prompt:"
Need to :
1. Move all the crud from @apiKeyService.ts to REST endpoints whcih can only be involed by auuthenticated endpoints. Make sure to implement also [id] endpoints
2. All operations should work for a user_id. You can get the users id by retrieving email from JWT and use that to get user id from supabase database

Ensure to use @supabaseApiKeyService.ts for database interactions.
Use @NextJS

"

'CRUD' Notepad contents that can be used as common context/prompt for this task.
"
Implement a CRUD for managing API keys. Original code is present in @apiKeyService.ts
"